Implement fast doubling algorithm
To improve the performance of calculating Fibonacci numbers, I implemented the fast doubling algorithm.

In addition, I used the clz (count leading zeros) instruction to count the input number's leading 0's bit, further optimizing the calculation.

As some systems may not support the clz instruction, I also added a version of the code that does not use it. This can be found in the comment section of the code.