Implement the big number API
To address the issue of overflow when calculating the Fibonacci number, I have introduced the concept of big numbers. This involves breaking down the number into multiple unsigned integers and developing essential operations to support big numbers, such as bn_add, bn_sub, bn_mul, and others.

With the inclusion of the bignum APIs, the Fibonacci device driver can now accurately calculate Fibonacci numbers up to at least the 3000th term (subject to available memory). The bignum APIs are defined in bn.h, and the corresponding source code can be found in bn.c.

Moreover, the read file operation to /dev/fibonacci has been updated to read the nth Fibonacci number into the user buffer, represented as a decimal string. It also returns the number of bytes that could not be copied.